# backend/scripts/swagger_v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DataLabellingSupportSystem.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.RegisterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.RegisterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login-google": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.GoogleLoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.GoogleLoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.GoogleLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.LogoutRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.LogoutRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.LogoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh-token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.RefreshTokenRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.RefreshTokenRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ForgotPasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ForgotPasswordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.ForgotPasswordResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.ForgotPasswordResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.ForgotPasswordResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ResetPasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ResetPasswordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ChangePasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ChangePasswordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/error_type": {
      "get": {
        "tags": [
          "ErrorType"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "ErrorType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateErrorTypeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateErrorTypeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateErrorTypeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "ErrorType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateErrorTypeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateErrorTypeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateErrorTypeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "ErrorType"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/error_type/{id}": {
      "get": {
        "tags": [
          "ErrorType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/review_errors/type/{id}": {
      "get": {
        "tags": [
          "ReviewErrors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_DataLabellingSupportSystem.Api.DTOs.Responses.Auth.ForgotPasswordResponse": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Auth.ForgotPasswordResponse"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.Common.Results.ServiceResponse_System.Boolean": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ForgotPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Requests.Auth.GoogleLoginRequest": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Requests.Auth.LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Requests.Auth.LogoutRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Requests.Auth.RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Requests.Auth.RegisterRequest": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "identifyNumber": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Requests.Auth.ResetPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "resetToken": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Requests.Reviews.UpdateErrorTypeRequest": {
        "type": "object",
        "properties": {
          "errorName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataLabellingSupportSystem.Api.DTOs.Responses.Auth.AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/DataLabellingSupportSystem.Api.DTOs.Responses.Auth.UserProfileResponse"
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Responses.Auth.ForgotPasswordResponse": {
        "type": "object",
        "properties": {
          "resetToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DataLabellingSupportSystem.Api.DTOs.Responses.Auth.UserProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "roleId": {
            "type": "string",
            "nullable": true
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste only the JWT access token here. Swagger will send: Authorization: Bearer <token>",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}